+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryScreenUi.kt
@@ -15,4 +15,5 @@ interface PatientSummaryScreenUi {
   fun showPatientDiedStatus()
   fun showNextAppointmentCard()
   fun hideNextAppointmentCard()
+  fun showClinicalDecisionSupportAlert()
 }